Wrap all the various way of doing checkouts in GTCheckoutOptions
Fixes #457
