+++ ObjectiveGit/ObjectiveGit.h
@@ -71,6 +71,7 @@ FOUNDATION_EXPORT const unsigned char ObjectiveGitVersionString[];
 #import <ObjectiveGit/GTFilterSource.h>
 #import <ObjectiveGit/GTFetchHeadEntry.h>
 #import <ObjectiveGit/GTNote.h>
+#import <ObjectiveGit/GTCheckoutOptions.h>
 
 #import <ObjectiveGit/GTObjectDatabase.h>
 #import <ObjectiveGit/GTOdbObject.h>
